% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.CLogitForest.R
\name{bootci.CLogitForest}
\alias{bootci.CLogitForest}
\title{Bootstrap confidence intervals for CLogitForest}
\usage{
\method{bootci}{CLogitForest}(model, B = 50, ncores = 1, alpha.ci = 0.05, quant.type = 7, ...)
}
\arguments{
\item{model}{Original CLogitForest model}

\item{B}{Number of bootstrap iterations}

\item{ncores}{Number of parallel cores to use}

\item{alpha.ci}{Confidence level of confidence intervals is calculated as \code{1-alpha.ci}.}

\item{quant.type}{Algorithm type used for quantiles, as described in argument \code{type} from \code{\link{quantile}}.}

\item{...}{Further bootci arguments}
}
\value{
\item{beta_hat}{Estimate for separate exposure  effect}
\item{ci_beta}{Estimated bootstrap confidence interval for exposure effect}
}
\description{
Performs estimation of bootstrap confidence intervals for separate exposure variable effects from \code{\link{CLogitForest}} objects.
}
\examples{
\donttest{
data(illu.small)

set.seed(1860)
illu.rf <- CLogitForest(illu.small, response = "y", exposure = "x", s = "strata",
                        ntree = 4, depth_max=2, tune.mtry = FALSE)

set.seed(1860)
rf.boot <- bootci(illu.rf, B = 10)
rf.boot
}
}
\seealso{
\code{\link{CLogitForest}}
}
\author{
Gunther Schauberger: \email{gunther.schauberger@tum.de} \cr
Moritz Berger: \email{moritz.berger@imbie.uni-bonn.de}
}
