% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.CLogitTree.R
\name{bootci.CLogitTree}
\alias{bootci.CLogitTree}
\title{Bootstrap confidence intervals for CLogitTree}
\usage{
\method{bootci}{CLogitTree}(model, B = 100, alpha.ci = 0.05, ...)
}
\arguments{
\item{model}{Original CLogitTree model}

\item{B}{Number of bootstrap iterations}

\item{alpha.ci}{Confidence level of confidence intervals is calculated as \code{1-alpha.ci}.}

\item{...}{Further bootci arguments}
}
\value{
\item{beta_hat}{Estimate for separate exposure  effect}
\item{ci_beta}{Estimated bootstrap confidence interval for exposure effect}
}
\description{
Performs estimation of bootstrap confidence intervals for separate exposure variable effects from \code{\link{CLogitTree}} objects.
}
\examples{
\donttest{
data(illu.small)

set.seed(1860)
illu.tree <- CLogitTree(illu.small, response = "y", exposure = "x", s = "strata",
                        alpha = 0.05, nperm = 20, trace = FALSE)

plot(illu.tree)

set.seed(1860)
illu.boot <- bootci(illu.tree, B = 10, alpha = 0.2)
}
}
\seealso{
\code{\link{CLogitTree}}, \code{\link{plot.CLogitTree}}, \code{\link{prune.CLogitTree}}
}
\author{
Gunther Schauberger: \email{gunther.schauberger@tum.de} \cr
Moritz Berger: \email{moritz.berger@imbie.uni-bonn.de}
}
