% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.CLogitTree.R
\name{bootci}
\alias{bootci}
\title{Bootstrap confidence intervals for CLogitTree}
\usage{
bootci(model, ...)
}
\arguments{
\item{model}{Object to calculate bootstrap confidence intervals for}

\item{...}{Further arguments}
}
\description{
Bootstrap confidence intervals for CLogitTree
}
