% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLogitForest.R
\name{CLogitForest}
\alias{CLogitForest}
\title{Function to perform CLogitForest}
\usage{
CLogitForest(
  data,
  response,
  exposure = NULL,
  s,
  ntree = 100,
  mtry = ifelse(is.null(exposure), max(floor((ncol(data) - 2)/3), 1),
    max(floor((ncol(data) - 3)/3), 1)),
  subsample.632 = FALSE,
  minnodesize = 15,
  minbucket = 5,
  depth_max = NULL,
  lambda = 1e-20,
  print.trace = FALSE,
  fit = TRUE,
  ncores = 2,
  BIC = FALSE,
  linear.offset = FALSE,
  tune.mtry = TRUE,
  tune.ntree = 20,
  tune.range = 2:floor((ncol(data) - 2) * 0.7),
  epsilon = 1e-05
)
}
\arguments{
\item{data}{Data frame (containing all required variable)}

\item{response}{Name of response variable (character)}

\item{exposure}{Name of (optional) separate exposure variable (character)}

\item{s}{Name of strata variable (character)}

\item{ntree}{Number of trees to fit for random forest}

\item{mtry}{Number of variables that a randomly selected for potential splits. Only effective if \code{tune.mtry = FALSE}}

\item{subsample.632}{If \code{TRUE}, only a subsample of 63.2% of all strata is
randomly selected for a single tree}

\item{minnodesize}{Minimal node size in order to be eligible for further splitting}

\item{minbucket}{Minimum number of observations in any terminal node}

\item{depth_max}{Maximum depth of the tree, with the root node counted as depth 0. If \code{NULL} (default), the size of the trees is not restricted.}

\item{lambda}{Tuning parameter for optional L2 penalty}

\item{print.trace}{Shall trace of permutation tests be printed?}

\item{fit}{Shall the internally fitted models be returned (required for the use of \code{\link[CLogitTree]{prune}})?}

\item{ncores}{Number of parallel nodes to use.}

\item{BIC}{Shall each tree be pruned using the BIC criterion?}

\item{linear.offset}{Shall a linear offset (from regular CLR) be used?}

\item{tune.mtry}{Shall \code{mtry} be tuned prior to model fitting? If true, \code{tune.ntree} trees are fitted with all possible
values of \code{mtry}}

\item{tune.ntree}{Number of trees to be used for \code{mtry} tuning. Only effective if \code{tune.mtry = TRUE}.}

\item{tune.range}{Range of possible values for mtry to be considered mtry-tuning. Only effective if \code{tune.mtry = TRUE}.}

\item{epsilon}{Convergence tolerance. Iteration continues until the relative change
in the conditional log likelihood is less than eps. Must be positive.}
}
\value{
\item{beta_hat}{Estimate for separate exposure  effect (arithmetic mean across exposure effects of all trees)}
\item{beta_median_hat}{Median across exposure effects of all trees}
\item{beta_list}{Vector containing all exposure estimates}
\item{tree_list}{List containing all CLogitTree objects which constitute the tree}
\item{mtry}{mtry argument from call}
\item{call}{function call}
\item{ncores}{Number of parallel nodes}
\item{minnodesize}{Minimal node size in order to be eligible for further splitting}
\item{minbucket}{Minimum number of observations in any terminal node}
\item{depth_max}{Maximum depth of the tree}
\item{lambda}{L2 tuning parameter}
\item{print.trace}{trace argument from call}
\item{subsample.632}{If \code{TRUE}, only a subsample of 63.2% of all strata was
randomly selected for a single tree}
\item{data}{Original data set}
\item{response}{Name of response variable}
\item{exposure}{Name of exposure variable}
\item{s}{Name of strata variable}
\item{ntree}{Number of trees to fit for random forest}
\item{BIC}{BIC argument from call}
\item{lin.off.model}{CLR model fitted for linear offset (if \code{linear.offset = TRUE})}
\item{offset.model}{If \code{linear.offset = TRUE}, this is the finally fitted model with the random forest as its offset.}
\item{epsilon}{Convergence tolerance}
}
\description{
Performs CLogitForest, a random forest method for the analysis of matched case-control studies. The estimation is embedded into the framework of conditional logistic regression.
Splits are incorporated into the model as separate dummy variables specifying the terminal nodes. Additional methods for pruning, plotting and bootstrap are available.
}
\examples{
data(illu.small)

set.seed(1860)
illu.rf <- CLogitForest(illu.small, response = "y", exposure = "x", s = "strata",
                        ntree = 4, depth_max = 2, tune.mtry = FALSE)

illu.rf
}
\seealso{
\code{\link{bootci.CLogitForest}}, \code{\link{predict.CLogitForest}}, \code{\link{varimp.CLogitForest}}
}
\author{
Gunther Schauberger: \email{gunther.schauberger@tum.de} \cr
Moritz Berger: \email{Moritz.Berger@imbie.uni-bonn.de}
}
