% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLogitTree.R
\name{CLogitTree}
\alias{CLogitTree}
\title{Function to perform CLogitTree}
\usage{
CLogitTree(
  data,
  response,
  exposure = NULL,
  s,
  alpha = 0.05,
  nperm = 500,
  minnodesize = 10,
  minbucket = 5,
  depth_max = NULL,
  perm_test = TRUE,
  mtry = NULL,
  lambda = 0,
  offset = NULL,
  print.trace = FALSE,
  ncores = 1,
  fit = TRUE,
  epsilon = 0.001
)
}
\arguments{
\item{data}{Data frame (containing all required variable)}

\item{response}{Name of response variable (character)}

\item{exposure}{Name of (optional) separate exposure variable (character)}

\item{s}{Name of strata variable (character)}

\item{alpha}{Level of significance for internal permutation tests}

\item{nperm}{Number of permutations in internal permutation tests}

\item{minnodesize}{Minimal node size in order to be eligible for further splitting}

\item{minbucket}{Minimum number of observations in any terminal node}

\item{depth_max}{Maximum depth of the tree, with the root node counted as depth 0. If \code{NULL} (default), the size of the trees is not restricted.}

\item{perm_test}{\code{TRUE} for regular use where  permutation tests are done.}

\item{mtry}{Solely for internal use}

\item{lambda}{Tuning parameter for optional L2 penalty}

\item{offset}{Optional offset used for model fitting}

\item{print.trace}{Shall trace of permutation tests be printed?}

\item{ncores}{Number of parallel nodes to use for permutation tests.}

\item{fit}{Shall the internally fitted models be returned (required for the use of \code{\link[CLogitTree]{prune}})?}

\item{epsilon}{Convergence tolerance. Iteration continues until the relative change
in the conditional log likelihood is less than eps. Must be positive.}
}
\value{
\item{beta_hat}{Estimate for separate exposure  effect}
\item{gamma_hat}{Estimates for terminal nodes  (with reference node)}
\item{gamma_hat_sym}{Estimates for terminal nodes (with symmetric side constraint)}
\item{splits}{Detailed information about all executed splits during the fitting process (matrix)}
\item{Z}{Design matrix of all confounding variables}
\item{y}{Outcome variable}
\item{y_tab}{Table of the number of cases and the number of controls in each node}
\item{strata}{Strata variable}
\item{exposure}{Exposure variable}
\item{model}{Models fitted internally}
\item{design}{Internally built design matrix including all indicator variables}
\item{param}{Fitted coefficients (solely for internal use)}
\item{param_fit}{Fitted coefficients (solely for internal use)}
\item{pvalue}{P-values of the permutation tests performed during the fitting process}
\item{dev}{Maximal value statistics of the selected variables in each iteration during the fitting process}
\item{crit}{Critical values of the permutation tests during the fitting process}
\item{BIC}{BICs of the internally fitted models (based on the unpenalized likelihood)}
\item{minnodesize}{Minimal node size in order to be eligible for further splitting}
\item{minbucket}{Minimum number of observations in any terminal node}
\item{depth_max}{Maximum depth of the tree}
\item{lambda}{L2 tuning parameter}
\item{print.trace}{print.trace argument from call}
\item{nperm}{nperm argument from call}
\item{alpha}{Level of significance for permutation tests}
\item{perm_test}{perm_test argument from call}
\item{mtry}{mtry argument from call}
\item{ncores}{Number of parallel nodes to use for permutation tests.}
\item{call}{function call}
\item{prunedBIC}{For internal use}
\item{epsilon}{Convergence tolerance}
}
\description{
Performs CLogitTree, a tree-based method for the analysis of matched case-control studies. The estimation is embedded into the framework of conditional logistic regression.
Splits are incorporated into the model as separate dummy variables specifying the terminal nodes. Additional methods for pruning, plotting and bootstrap are available.
}
\examples{
data(illu.small)

set.seed(1860)
illu.tree <- CLogitTree(illu.small, response = "y", exposure = "x", s = "strata",
                        alpha = 0.05, nperm = 20)

plot(illu.tree)
## preferred procedure: pruning using BIC
set.seed(1860)
illu.tree <- CLogitTree(illu.small, response = "y", exposure = "x", s = "strata",
                        perm_test = FALSE, depth_max=4)
illu.tree <- pruneBIC(illu.tree)
}
\seealso{
\code{\link{plot.CLogitTree}}, \code{\link{bootci.CLogitTree}}, \code{\link{prune.CLogitTree}}, \code{\link{predict.CLogitTree}}
}
\author{
Gunther Schauberger: \email{gunther.schauberger@tum.de} \cr
Moritz Berger: \email{Moritz.Berger@imbie.uni-bonn.de}
}
