% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLogitTree.R
\docType{data}
\name{illu.data}
\alias{illu.data}
\title{Illustrative data}
\format{
A data frame with 1600 rows and 8 variables
\describe{
  \item{x}{Exposure variable}
  \item{strata}{Strata variable}#'
  \item{Z1}{Explanatory/confounder variable}
  \item{Z2}{Explanatory/confounder variable}
  \item{Z3}{Explanatory/confounder variable}
  \item{Z4}{Explanatory/confounder variable}
  \item{Z5}{Explanatory/confounder variable}
  \item{y}{Case-control status (1: case; 0: control)}
}
}
\description{
A simulated dataset for illustration of the CLogitTree method.
}
