% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLogitTree.R
\docType{data}
\name{illu.small}
\alias{illu.small}
\title{Illustrative data}
\format{
A data frame with 400 rows and 5 variables
\describe{
  \item{x}{Exposure variable}
  \item{strata}{Strata variable}
  \item{Z4}{Explanatory/confounder variable}
  \item{Z5}{Explanatory/confounder variable}
  \item{y}{Case-control status (1: case; 0: control)}
}
}
\description{
A small simulated dataset for illustration of the CLogitTree method.
}
