% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CLogitTree.R
\name{plot.CLogitTree}
\alias{plot.CLogitTree}
\title{Plot function for CLogitTree}
\usage{
\method{plot}{CLogitTree}(
  x,
  symmetric = TRUE,
  precision = 3,
  ellipse_a = 0.8,
  ellipse_b = 0.2,
  ellipse_x = 0,
  ellipse_y = 0.2,
  branch_adj = 0,
  cex.lines = 2,
  cex.branches = 1,
  cex.coefs = 1,
  cex.main = 1,
  cex.numbers = 1,
  info_inner = TRUE,
  info_n = TRUE,
  draw_numbers = FALSE,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{CLogitTree object}

\item{symmetric}{Shall parameter estimates with symmetric side constraints be used?}

\item{precision}{Number of decimals of parameter estimates}

\item{ellipse_a}{Controls width of ellipse containing node information}

\item{ellipse_b}{Controls height of ellipse containing node information}

\item{ellipse_x}{Controls location on x-axis of ellipse containing node information}

\item{ellipse_y}{Controls location on y-axis of ellipse containing node information}

\item{branch_adj}{Vertical adjustment of branch labels}

\item{cex.lines}{Size of lines}

\item{cex.branches}{Size of branch labels}

\item{cex.coefs}{Size of parameter estimates}

\item{cex.main}{Size of title}

\item{cex.numbers}{Size of numbers of observations}

\item{info_inner}{Shall +/- signs be added to split points?}

\item{info_n}{Shall numbers of observations be displayed?}

\item{draw_numbers}{Shall internally used node numbers be displayed?}

\item{title}{Main title}

\item{...}{further plot arguments}
}
\value{
No return, used for side effects
}
\description{
Plots trees estimated by \code{\link{CLogitTree}}.
}
\examples{
data(illu.small)

set.seed(1860)
illu.tree <- CLogitTree(illu.small, response = "y", exposure = "x", s = "strata",
                        alpha = 0.05, nperm = 20, print.trace = FALSE)

plot(illu.tree)

}
\seealso{
\code{\link{CLogitTree}}, \code{\link{bootci.CLogitTree}}, \code{\link{prune.CLogitTree}}
}
\author{
Gunther Schauberger: \email{gunther.schauberger@tum.de} \cr
Moritz Berger: \email{moritz.berger@imbie.uni-bonn.de}
}
