% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.varimp.R
\name{plot.varimp.CLF}
\alias{plot.varimp.CLF}
\title{Plot function for variable importance of \code{\link{CLogitForest}}}
\usage{
\method{plot}{varimp.CLF}(
  x,
  sort = TRUE,
  horiz = TRUE,
  las = 1,
  col = "deepskyblue4",
  border = FALSE,
  ...
)
}
\arguments{
\item{x}{object containing variable importance results for a \code{\link{CLogitForest}} object}

\item{sort}{Shall the variables be plotted sorted according to their variable importance values?}

\item{horiz}{See \code{\link{barplot}}}

\item{las}{See \code{\link{barplot}}}

\item{col}{See \code{\link{barplot}}}

\item{border}{See \code{\link{barplot}}}

\item{...}{further \code{\link{barplot}} arguments}
}
\value{
No return, used for side effects
}
\description{
Plots barplots which represent the variable importance in \code{\link{CLogitForest}} objects calculated via \code{\link{varimp.CLogitForest}}.
}
\examples{
data(illu.small)

}
\seealso{
\code{\link{CLogitForest}}, \code{\link{varimp.CLogitForest}}
}
\author{
Gunther Schauberger: \email{gunther.schauberger@tum.de} \cr
Moritz Berger: \email{moritz.berger@imbie.uni-bonn.de}
}
