% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.CLogitForest.R
\name{predict.CLogitForest}
\alias{predict.CLogitForest}
\title{Predict function for CLogitForest}
\usage{
\method{predict}{CLogitForest}(
  object,
  response = NULL,
  exposure = NULL,
  s = NULL,
  newdata = NULL,
  ncores = 1,
  type = c("link", "response", "loglik", "forest"),
  median.exposure = FALSE,
  ...
)
}
\arguments{
\item{object}{CLogitForest object}

\item{response}{Name of response variable (character)}

\item{exposure}{Name of (optional) separate exposure variable (character)}

\item{s}{Name of strata variable (character)}

\item{newdata}{(Optional) new data frame. If \code{newdata = NULL}, the original data set from \code{object} is used.}

\item{ncores}{Number of parallel cores to use}

\item{type}{Type of prediction that is required. For option \code{"link"} (default), the prediction of the
linear predictor per observation is returned. For option \code{"response"} (default), the prediction of the
probability per observation is returned. This probability is conditional on the stratum,
i.e. the probabilities per stratum add up to 1. For option \code{"loglik"}, the prediction of the conditional log-likelihood
per stratum is returned. For option \code{"forest"}, the prediction of the random forest (without exposure effect)
per observation is returned.}

\item{median.exposure}{Shall the median of the exposure effects from single trees be used as
estimate (instead of arithmetic mean)}

\item{...}{Further predict arguments}
}
\value{
Vector containing predicted linear predictors
}
\description{
Predicts the linear predictor of estimated CLogitForest models
}
\seealso{
\code{\link{CLogitForest}}
}
\author{
Gunther Schauberger: \email{gunther.schauberger@tum.de} \cr
Moritz Berger: \email{moritz.berger@imbie.uni-bonn.de}
}
