% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.CLogitTree.R
\name{predict.CLogitTree}
\alias{predict.CLogitTree}
\title{Predict for CLogitTree}
\usage{
\method{predict}{CLogitTree}(object, response, exposure = NULL, s, newdata = NULL, ...)
}
\description{
Predict for CLogitTree
}
