% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.CLogitTree.R
\name{prune.CLogitTree}
\alias{prune.CLogitTree}
\title{Pruning for CLogitTree}
\usage{
\method{prune}{CLogitTree}(tree, alpha, ...)
}
\arguments{
\item{tree}{Original CLogitTree model}

\item{alpha}{Level of significance used in internal permutation tests}

\item{...}{Further prune arguments}
}
\value{
\code{\link{CLogitTree}} object
}
\description{
Prunes trees built by \code{\link{CLogitTree}}. Pruning is steered by the significance level used within permutation tests.
}
\examples{
data(illu.small)

set.seed(1860)
illu.tree <- CLogitTree(illu.small, response = "y", exposure = "x", s = "strata",
                        alpha = 0.05, nperm = 20, trace = FALSE)

plot(prune(illu.tree,0.4))

}
\seealso{
\code{\link{CLogitTree}}, \code{\link{plot.CLogitTree}}, \code{\link{bootci.CLogitTree}}
}
\author{
Gunther Schauberger: \email{gunther.schauberger@tum.de} \cr
Moritz Berger: \email{moritz.berger@imbie.uni-bonn.de}
}
