% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneBIC.CLogitTree.R
\name{pruneBIC.CLogitTree}
\alias{pruneBIC.CLogitTree}
\title{Pruning for CLogitTree}
\usage{
\method{pruneBIC}{CLogitTree}(tree, ...)
}
\arguments{
\item{tree}{Original CLogitTree model}

\item{...}{Further prune arguments}
}
\value{
\code{\link{CLogitTree}} object
}
\description{
Prunes trees built by \code{\link{CLogitTree}}. Pruning is steered by the BIC of the sequence of fitted models during tree building.
}
\examples{
data(illu.small)

set.seed(1860)
illu.tree <- CLogitTree(illu.small, response = "y", exposure = "x", s = "strata",
                        alpha = 0.05, nperm = 20, print.trace = FALSE)

plot(pruneBIC(illu.tree))

}
\seealso{
\code{\link{CLogitTree}}, \code{\link{plot.CLogitTree}}, \code{\link{bootci.CLogitTree}}
}
\author{
Gunther Schauberger: \email{gunther.schauberger@tum.de} \cr
Moritz Berger: \email{moritz.berger@imbie.uni-bonn.de}
}
