% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneBIC.CLogitTree.R
\name{pruneBIC}
\alias{pruneBIC}
\title{Pruning for CLogitTree}
\usage{
pruneBIC(tree, ...)
}
\arguments{
\item{tree}{Original CLogitTree model}

\item{...}{Further arguments}
}
\description{
Pruning for CLogitTree
}
