% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp.R
\name{varimp}
\alias{varimp}
\title{Variable importance for CLogitForest}
\usage{
varimp(model, ...)
}
\arguments{
\item{model}{Object to calculate variable importance for}

\item{...}{Further arguments}
}
\description{
Variable importance for CLogitForest
}
